Move projects header to app component
Moving header of projects component to app component will allow a global
state holder. This will make it possible to keep track of the selections
and states and also removes duplicate code. Also some optimizations and
logical actions that require the states of multiple components will be
easier to implement.

IMPORTANT: This commit does not do any changes on project view component.
This will follow in a future change adding more optimizations and removal
of duplicate code.